+++ src/page-layout/NoticesLayout/components/AllNotices/AllNotices.module.scss
@@ -1,49 +1,1 @@
-.allNotices {
-  max-width: 96.4rem;
-  background-color: $color-black;
-}
 
-.header {
-  position: relative;
-  width: 100%;
-  margin-bottom: 3.2rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  .heading {
-    flex-basis: 50%;
-    @include sopoqa-28-700;
-  }
-
-  .filterBar {
-    flex-basis: 50%;
-    text-align: right;
-  }
-
-  .filterItem {
-    display: inline;
-
-    .filterButton {
-      background-color: $color-red-30;
-      margin-left: 1rem;
-    }
-  }
-}
-
-.filter {
-  top: 4rem;
-  right: 0;
-}
-
-.postContainer {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  column-gap: 1.4rem;
-  row-gap: 3rem;
-}
-
-.post {
-  max-width: 31.2rem;
-  height: 35rem;
-}
