+++ src/page-layout/NoticesLayout/components/AllNoticesSkeleton/AllNoticesSkeleton.module.scss
@@ -1,0 +1,39 @@
+.header {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 3.1rem;
+
+  .heading {
+    width: 11.2rem;
+    height: 3.5rem;
+  }
+
+  .filter {
+    width: 15rem;
+    height: 3.5rem;
+  }
+}
+
+.list {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  column-gap: 1.4rem;
+  row-gap: 3rem;
+
+  .post {
+    width: 31.2rem;
+    height: 35rem;
+  }
+}
+
+@include tablet {
+  .list {
+    grid-template-columns: 1fr 1fr;
+  }
+}
+
+@include mobile {
+  .list {
+    grid-template-columns: 1fr;
+  }
+}
