+++ src/page-layout/NoticesLayout/components/AllNoticesSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AllNoticesSkeleton";
