+++ src/page-layout/NoticesLayout/components/FilterBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FilterBar";
