+++ src/page-layout/NoticesLayout/components/NoNotice/NoNotice.module.scss
@@ -1,0 +1,3 @@
+.container {
+  width: 100%;
+}
