+++ src/page-layout/NoticesLayout/components/NoNotice/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NoNotice";
