+++ src/page-layout/NoticesLayout/components/NoticeList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NoticeList";
