+++ src/page-layout/NoticesLayout/components/index.ts
@@ -1,1 +1,2 @@
-export { default } from "./AllNotices/AllNotices";
+export { default as AllNotices } from "./AllNotices/AllNotices";
+export { default as NoNotice } from "./NoNotice/NoNotice";
