+++ src/page-layout/NoticesLayout/type.ts
@@ -1,0 +1,8 @@
+type Sort = "time" | "pay" | "hour" | "shop" | "";
+
+export interface FilterValue {
+  address: string[];
+  startsAtGte: string;
+  hourlyPayGte: string;
+  sort: Sort;
+}
