+++ src/pages/notices/notices.module.scss
@@ -2,3 +2,8 @@
   display: flex;
   justify-content: center;
 }
+
+.allNotices {
+  max-width: 96.4rem;
+  margin: 6rem 0;
+}
