+++ src/shared/hooks/useGetAllNotices.ts
@@ -2,7 +2,7 @@ import { keepPreviousData, useQuery } from "@tanstack/react-query";
 import { axiosInstance } from "@/shared/apis/axiosInstance";
 import { CATEGORIES } from "@/common/constants";
 
-import type { FilterValue } from "@/pages/notices";
+import type { FilterValue } from "@/page-layout/NoticesLayout/type";
 
 type Category = (typeof CATEGORIES)[number];
 
@@ -50,9 +50,6 @@ interface NoticeResponse {
 const fetchNotices = async (offset: number, filters: Partial<FilterValue>): Promise<NoticeResponse> => {
   const res = await axiosInstance.get(`/notices`, {
     params: { offset, limit: 6, ...filters },
-    paramsSerializer: {
-      indexes: null,
-    },
   });
 
   const result = await res.data;
@@ -61,13 +58,13 @@ const fetchNotices = async (offset: number, filters: Partial<FilterValue>): Prom
 };
 
 const useGetAllNotices = (offset: number, filters: Partial<FilterValue>) => {
-  const { data, isLoading, error, isPending, isError } = useQuery({
+  const { data, error, isPending, isError } = useQuery({
     queryKey: ["notices", offset, filters],
     queryFn: () => fetchNotices(offset, filters),
     placeholderData: keepPreviousData,
   });
 
-  return { data, isLoading, error, isPending, isError };
+  return { data, error, isPending, isError };
 };
 
 export default useGetAllNotices;
